requirements.txt
langchain
Pillow
IPython
gigachat
python-telegram-bot[ext]
replicate
translate
shutil
requests
flask
flask-wtf